Sample precinct data from 2020 presidential election

A dataset containing precinct-by-precinct election results in Saratoga County, NY, for the 2020 presidential election. Note that New York's unusual rules allow candidates to be listed under more than one party on the ballot, which is why you'll see Trump in there twice under both the Republican and Conservative parties.

precinctsampledata_ny

Format

A data frame with 196 rows and 11 variables:

precinct

election precinct name

'Joseph R. Biden - DEM'

votes for Biden as Democrat

'Donald J. Trump - REP'

votes for Trump as Republican

'Donald J. Trump - CON'

votes for Trump as Conservative Party

'Howie Hawkins - GRE'

votes for Hawkins as Green Party

'Jo Jorgensen - LIB'

votes for Jogensen as Libertartian

'Brock Pierce - IND'

votes for Pierce as Independence Party

WriteIn

votes for write-in candidates

Blanks

undervotes, where voter skipped the race and left it blank

Voids

overvotes, where voter improperly marked multiple candidates

...

Source

https://www.saratogacountyny.gov/2020-official-general-election-results/